README.md: Wrap long lines and new sentences
Long lines of plain text can  be irritating to work with; an
editor  might  not  "soft"  wrap the  lines  at  a  suitable
boundary (or at all).

Moreover, diff and  patch programs (like `git  diff' or `git
apply',  etc.) are  line-based; if  a paragraph  of text  is
written as one line of text,  then changing a single word in
that paragraph  produces a diff  which not only  entails the
entire paragraph, but also which  duplicates the rest of the
paragraph.

The long-time  solution to these primitive  line-based tools
is to split long "logical" lines across multiple, relatively
short  "file"  lines;  furthermore, it  is  a  long-standing
convention in these  situations to start each  sentence on a
new file line,  so that each sentence  can be diffed/patched
relatively independently.

This commit  breaks long  logically contiguous  lines across
multiple, relatively  short file lines; only  white space in
the  source has  been changed,  which can  be seen  with the
following POSIX shell commands:

  #!/bin/sh

  # Grab a well-known copy of the file:

  git show >/tmp/original \
    5de337bfa018e480bf15b77973e27ccdbada8e56:README.md

  # Here, sed unwraps the README.md provided by this
  # commit; note that only white space is manipulated:

  cat README.md | sed -n >/tmp/new-unwrapped '
    H
    /^$/b print
    $!b
    : print
    x
    s/^\n//
    s/\([^[:space:]]\)[[:space:]]\{1,\}\([^*[:space:]]\)/\1 \2/g
    p
  '

  # This compares the original file to the new,
  # unwrapped version:

  git diff /tmp/original /tmp/new-unwrapped && echo No Change

Alternatively, a less robust test  is just to remove all the
white space from each version, and then do the comparison:

  #!/bin/sh

  sed_program='H;$!b;x;s/[[:space:]]//g;p'

  git show 5de337bfa018e480bf15b77973e27ccdbada8e56:README.md |
    sed -n "$sed_program" >/tmp/original-squashed
  cat README.md |
    sed -n "$sed_program" >/tmp/new-squashed

  git diff /tmp/original-squashed /tmp/new-squashed && echo No Change

Both programs should produce this output:

  No Change